Speedup job data serialization up to 750x.
The current implementation used to concatenate long strings to
generate the 'data' entry to be dispatched to the selected jobds.
For large deployments and churn traces, especially under the --trace_alt
mode, this results in big (several dozens of MB) that are generated.
This modification generates a simple Hash object that is filled
with the required fields; then, it is serialized by the 'json' native
library.
On our tests, this modification had major benefits, with a single
serialization from 30s to 0.02s on average.